requirements.txt
Django==2.2.19
djangorestframework==3.12.4
pytz==2022.2.1
sqlparse==0.4.2
